RemoteOnlyPolicy: able to run
